// src/components/map/types.ts
export interface Planets {
  sun: { url: string; size: number; dist: number };
  mercury: { url: string; size: number; dist: number };
  venus: { url: string; size: number; dist: number };
  earth: { url: string; size: number; dist: number };
  mars: { url: string; size: number; dist: number };
  jupiter: { url: string; size: number; dist: number };
  saturn: { url: string; size: number; dist: number };
  uranus: { url: string; size: number; dist: number };
  neptune: { url: string; size: number; dist: number };
  pluto: { url: string; size: number; dist: number };
}